pyproject.toml
[project]
name = "alder"
version = "0.1.0"
description = "Local-first ERP for a 3D printing business (Clean Architecture, SQLAlchemy 2.0 + Django UI)."
requires-python = ">=3.11"
dependencies = [
    "SQLAlchemy>=2.0.25",
    "Django>=5.0,<6.0",
    "pydantic>=2.5",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "pandas>=2.2",
    "openpyxl>=3.1",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.8",
]
postgres = ["psycopg[binary]>=3.1"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
